requirements.txt
Django==5.1.2
Pillow==11.0.0
black==24.10.0
celery-redbeat==2.2.0
celery[redis]==5.4.0
crispy-tailwind==1.0.3
django-allauth==65.1.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-post_office==3.9.0
django-ses==4.2.0
django-stubs[compatible-mypy]==5.1.0
djangoql==0.18.1
gunicorn==23.0.0
psycopg[binary]==3.2.3
pyjwt[crypto]==2.9.0
requests==2.32.3
whitenoise==6.7.0
